Restructured files and .gitignore to the new CLion IDE.
