--- .vs/VSWorkspaceState.json
@@ -1,8 +1,0 @@
-{
-  "ExpandedNodes": [
-    "",
-    "\\Development Files"
-  ],
-  "SelectedNode": "\\Development Files\\blackjack.h",
-  "PreviewInSolutionExplorer": false
-}
