--- .vs/console-casino.slnx/v18/DocumentLayout.backup.json
@@ -1,37 +1,0 @@
-{
-  "Version": 1,
-  "WorkspaceRootPath": "C:\\Users\\lilmc\\OneDrive\\Documents\\Repos\\console-casino\\",
-  "Documents": [
-    {
-      "AbsoluteMoniker": "D:0:0:{10CCA09A-43AE-4508-A401-99E88D074CAF}|console-casino.vcxproj|C:\\Users\\lilmc\\OneDrive\\Documents\\Repos\\console-casino\\Development Files\\main.cpp||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}",
-      "RelativeMoniker": "D:0:0:{10CCA09A-43AE-4508-A401-99E88D074CAF}|console-casino.vcxproj|solutionrelative:Development Files\\main.cpp||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}"
-    }
-  ],
-  "DocumentGroupContainers": [
-    {
-      "Orientation": 0,
-      "VerticalTabListWidth": 256,
-      "DocumentGroups": [
-        {
-          "DockedWidth": 200,
-          "SelectedChildIndex": 0,
-          "Children": [
-            {
-              "$type": "Document",
-              "DocumentIndex": 0,
-              "Title": "main.cpp",
-              "DocumentMoniker": "C:\\Users\\lilmc\\OneDrive\\Documents\\Repos\\console-casino\\Development Files\\main.cpp",
-              "RelativeDocumentMoniker": "Development Files\\main.cpp",
-              "ToolTip": "C:\\Users\\lilmc\\OneDrive\\Documents\\Repos\\console-casino\\Development Files\\main.cpp",
-              "RelativeToolTip": "Development Files\\main.cpp",
-              "ViewState": "AgIAAAAAAAAAAAAAAAAAAAUAAAAtAAAAAAAAAA==",
-              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000677|",
-              "WhenOpened": "2026-03-09T06:30:16.853Z",
-              "EditorCaption": ""
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}
