+++ CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 4.1)
+project(console_casino)
+
+set(CMAKE_CXX_STANDARD 20)
+
+add_executable(console_casino main.cpp)
