--- Headers/utilities.h
@@ -1,15 +1,0 @@
-#ifndef UTILITIES_H
-#define UTILITIES_H
-
-// Function Prototypes
-void clearConsole();
-void showChips();
-int getPlayerBet();
-int generateNewCard();
-
-// Variable Declarations
-extern int playerChips;
-extern int playerBet;
-extern int currentCard;
-
-#endif // !UTILITIES_H
