+++ utilities.h
@@ -1,0 +1,15 @@
+#ifndef UTILITIES_H
+#define UTILITIES_H
+
+// Function Prototypes
+void clearConsole();
+void showChips();
+int getPlayerBet();
+int generateNewCard();
+
+// Variable Declarations
+extern int playerChips;
+extern int playerBet;
+extern int currentCard;
+
+#endif // !UTILITIES_H
